chore: refactor to util
