+++ src/shared/hooks/useDaoPlugins/index.ts
@@ -1,5 +1,4 @@
 export {
-    compareFilterPlugins,
     type IUseDaoPluginsParams,
     pluginGroupFilter,
     useDaoPlugins,
