+++ src/shared/hooks/useDaoPlugins/useDaoPlugins.test.ts
@@ -1,5 +1,4 @@
 import { renderHook } from '@testing-library/react';
-import type { IDaoPlugin } from '@/shared/api/daoService';
 import * as daoService from '@/shared/api/daoService';
 import { PluginInterfaceType } from '@/shared/api/daoService';
 import { FeatureFlagsProvider } from '@/shared/components/featureFlagsProvider';
@@ -10,7 +9,7 @@ import {
 } from '@/shared/testUtils';
 import { PluginType } from '@/shared/types';
 import { daoUtils } from '@/shared/utils/daoUtils';
-import { compareFilterPlugins, useDaoPlugins } from './useDaoPlugins';
+import { useDaoPlugins } from './useDaoPlugins';
 
 jest.mock('@/shared/api/cmsService', () => ({
     useDaoOverrides: () => ({ data: undefined }),
@@ -112,178 +111,5 @@ describe('useDaoPlugins hook', () => {
             wrapper: FeatureFlagsProvider,
         });
         expect(getDaoPluginsSpy).toHaveBeenCalledWith(dao, { hasExecute });
-    });
-});
-
-describe('compareFilterPlugins', () => {
-    const ROOT_DAO_ADDRESS = '0xRoot';
-
-    const buildFilterPlugin = (plugin: Partial<IDaoPlugin>) => ({
-        id: plugin.interfaceType ?? PluginInterfaceType.UNKNOWN,
-        uniqueId: `${plugin.address ?? '0x0'}-${plugin.slug ?? 'slug'}`,
-        label: '',
-        meta: generateDaoPlugin(plugin),
-        props: {},
-    });
-
-    it('sorts root-DAO plugins before sub-plugins of the same type', () => {
-        const root = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.MULTISIG,
-            daoAddress: ROOT_DAO_ADDRESS,
-            isSubPlugin: false,
-        });
-        const sub = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.MULTISIG,
-            daoAddress: ROOT_DAO_ADDRESS,
-            isSubPlugin: true,
-        });
-
-        const sorted = [sub, root].sort((a, b) =>
-            compareFilterPlugins(a, b, ROOT_DAO_ADDRESS),
-        );
-        expect(sorted).toEqual([root, sub]);
-    });
-
-    it('sorts root-DAO plugins before linked-account plugins of the same type', () => {
-        const root = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.TOKEN_VOTING,
-            daoAddress: ROOT_DAO_ADDRESS,
-            isSubPlugin: false,
-        });
-        const linked = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.TOKEN_VOTING,
-            daoAddress: '0xLinkedDao',
-            isSubPlugin: false,
-        });
-
-        const sorted = [linked, root].sort((a, b) =>
-            compareFilterPlugins(a, b, ROOT_DAO_ADDRESS),
-        );
-        expect(sorted).toEqual([root, linked]);
-    });
-
-    it('sorts by interface type priority: TOKEN_VOTING > MULTISIG > LOCK_TO_VOTE', () => {
-        const lock = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.LOCK_TO_VOTE,
-            daoAddress: ROOT_DAO_ADDRESS,
-            address: '0x3',
-        });
-        const multisig = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.MULTISIG,
-            daoAddress: ROOT_DAO_ADDRESS,
-            address: '0x2',
-        });
-        const token = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.TOKEN_VOTING,
-            daoAddress: ROOT_DAO_ADDRESS,
-            address: '0x1',
-        });
-
-        const sorted = [lock, multisig, token].sort((a, b) =>
-            compareFilterPlugins(a, b, ROOT_DAO_ADDRESS),
-        );
-        expect(sorted.map((p) => p.meta.interfaceType)).toEqual([
-            PluginInterfaceType.TOKEN_VOTING,
-            PluginInterfaceType.MULTISIG,
-            PluginInterfaceType.LOCK_TO_VOTE,
-        ]);
-    });
-
-    it('sorts unknown/other interface types after listed types with priority 99', () => {
-        const admin = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.ADMIN,
-            daoAddress: ROOT_DAO_ADDRESS,
-            address: '0x1',
-        });
-        const lockToVote = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.LOCK_TO_VOTE,
-            daoAddress: ROOT_DAO_ADDRESS,
-            address: '0x2',
-        });
-
-        const sorted = [admin, lockToVote].sort((a, b) =>
-            compareFilterPlugins(a, b, ROOT_DAO_ADDRESS),
-        );
-        expect(sorted.map((p) => p.meta.interfaceType)).toEqual([
-            PluginInterfaceType.LOCK_TO_VOTE,
-            PluginInterfaceType.ADMIN,
-        ]);
-    });
-
-    it('treats plugins with undefined daoAddress as non-root without throwing', () => {
-        const root = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.MULTISIG,
-            daoAddress: ROOT_DAO_ADDRESS,
-            isSubPlugin: false,
-            address: '0x1',
-        });
-        const undefinedAddress = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.MULTISIG,
-            isSubPlugin: false,
-            address: '0x2',
-        });
-
-        const sorted = [undefinedAddress, root].sort((a, b) =>
-            compareFilterPlugins(a, b, ROOT_DAO_ADDRESS),
-        );
-        expect(sorted).toEqual([root, undefinedAddress]);
-    });
-
-    it('preserves original order for equal-priority plugins (stable sort)', () => {
-        const first = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.MULTISIG,
-            daoAddress: ROOT_DAO_ADDRESS,
-            address: '0xFirst',
-        });
-        const second = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.MULTISIG,
-            daoAddress: ROOT_DAO_ADDRESS,
-            address: '0xSecond',
-        });
-
-        const sorted = [first, second].sort((a, b) =>
-            compareFilterPlugins(a, b, ROOT_DAO_ADDRESS),
-        );
-        expect(sorted.map((p) => p.meta.address)).toEqual([
-            '0xFirst',
-            '0xSecond',
-        ]);
-    });
-
-    it('sorts a mixed set: root by type, then non-root by type', () => {
-        const rootToken = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.TOKEN_VOTING,
-            daoAddress: ROOT_DAO_ADDRESS,
-            isSubPlugin: false,
-            address: '0x1',
-        });
-        const rootMultisig = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.MULTISIG,
-            daoAddress: ROOT_DAO_ADDRESS,
-            isSubPlugin: false,
-            address: '0x2',
-        });
-        const linkedMultisig = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.MULTISIG,
-            daoAddress: '0xLinked',
-            isSubPlugin: false,
-            address: '0x3',
-        });
-        const subLock = buildFilterPlugin({
-            interfaceType: PluginInterfaceType.LOCK_TO_VOTE,
-            daoAddress: ROOT_DAO_ADDRESS,
-            isSubPlugin: true,
-            address: '0x4',
-        });
-
-        const sorted = [subLock, linkedMultisig, rootMultisig, rootToken].sort(
-            (a, b) => compareFilterPlugins(a, b, ROOT_DAO_ADDRESS),
-        );
-        expect(sorted.map((p) => p.meta.address)).toEqual([
-            '0x1',
-            '0x2',
-            '0x3',
-            '0x4',
-        ]);
     });
 });
