+++ src/shared/utils/pluginSortUtils/index.ts
@@ -1,0 +1,5 @@
+export {
+    defaultTypePriority,
+    type ISortPluginsByDisplayOrderParams,
+    pluginSortUtils,
+} from './pluginSortUtils';
